requirements.txt
# No external dependencies required - using built-in Tkinter 
customtkinter
pillow